// types/house.ts
export interface House {
  access: string;
  accommodates: number;
  amenities: string[];
  availability_30: number;
  availability_365: number;
  availability_60: number;
  availability_90: number;
  bathrooms: number;
  bed_type: string;
  bedrooms: number;
  beds: number;
  calculated_host_listings_count: number;
  calendar_last_scraped: Date;
  calendar_updated: string;
  cancellation_policy: string;
  city: string;
  cleaning_fee: null;
  country: string;
  country_code: string;
  description: string;
  experiences_offered: string;
  extra_people: number;
  features: string[];
  first_review: Date;
  geolocation: Geolocation;
  guests_included: number;
  has_availability: null;
  host_about: null;
  host_acceptance_rate: string;
  host_id: string;
  host_listings_count: number;
  host_location: string;
  host_name: string;
  host_neighbourhood: null;
  host_picture_url: string;
  host_response_rate: number;
  host_response_time: string;
  host_since: Date;
  host_thumbnail_url: string;
  host_total_listings_count: number;
  host_url: string;
  host_verifications: string[];
  house_rules: string;
  id: string;
  interaction: string;
  jurisdiction_names: null;
  last_review: Date;
  last_scraped: Date;
  latitude: string;
  license: null;
  listing_url: string;
  longitude: string;
  market: string;
  maximum_nights: number;
  medium_url: string;
  minimum_nights: number;
  monthly_price: null;
  name: string;
  neighborhood_overview: null;
  neighbourhood: null;
  neighbourhood_cleansed: string;
  neighbourhood_group_cleansed: null;
  notes: null;
  number_of_reviews: number;
  picture_url: null;
  price: number;
  property_type: string;
  review_scores_accuracy: number;
  review_scores_checkin: number;
  review_scores_cleanliness: number;
  review_scores_communication: number;
  review_scores_location: number;
  review_scores_rating: number;
  review_scores_value: number;
  reviews_per_month: number;
  room_type: string;
  scrape_id: string;
  security_deposit: null;
  smart_location: string;
  space: null;
  square_feet: null;
  state: string;
  street: string;
  summary: string;
  thumbnail_url: string;
  transit: null;
  weekly_price: null;
  xl_picture_url: string;
  zipcode: string;
}

export interface Geolocation {
  lon: number;
  lat: number;
}